+++ Brocken_Physik/Sphere.cpp
@@ -4,7 +4,7 @@
 
 
 
-void Sphere::set(const Vector3f& pos, const f32& r, Material k, const f32& m){
+void Sphere::set(const Vector3d& pos, const f32& r, Material k, const f32& m){
 	x = pos;
 	this->r = r;
 	this->k = k;
